Add feature - admin functionality & user controler & views restrictions

+++ app/controllers/admin/users_controller.rb
@@ -1,15 +1,21 @@
 module Admin
   class UsersController < AdminController
+
     def show
       @user = User.find(params[:id])
+      authorize @user
     end
 
     def new
       @user = User.new
+      authorize @user
     end
 
     def edit
       @user = User.find(params[:id])
+
+      authorize @user
+
     end
 
     def index
@@ -18,22 +24,25 @@ module Admin
 
     def create
       @user = User.new(sign_params)
+      authorize @user
       if @user.save
         session[:user_id] = @user.id
         redirect_to admin_user_path(@user)
       else
-        flash[:error] = "Invalid something"
+
         render new_admin_user_path
       end
     end
 
     def destroy
       @user = User.find(params[:id])
+      authorize @user
       redirect_to admin_users_path(@user) if @user.destroy
     end
 
     def update
       @user = User.find(params[:id])
+
       if @user.update(sign_params)
         redirect_to admin_user_path(@user)
       else
@@ -46,6 +55,7 @@ module Admin
 
     def sign_params
       params.require(:user).permit(:name, :email, :password, :password_confirmation, :phone)
+
     end
   end
 end
